requirements.txt
Django==3.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
psycopg2==2.9.5
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0

